add: Comments adapter
